requirements.txt
langchain
langchain_core
langchain-community
langchain-groq
groq



pydantic
# environment varilables
python-dotenv

GitPython
# cli handling
click